refactor(tests): Move mock storage implementation to it's own package
